refactor(models): extract Transformer.InstanceRef module per spec package (Rule 0007 A)
